+++ src/cpp/flann/util/result_set.h
@@ -177,7 +177,7 @@ public:
      * @param num_elements Number of elements to copy
      * @param sorted Indicates if results should be sorted
      */
-    void copy(int* indices, DistanceType* dists, size_t num_elements, bool sorted = true)
+    void copy(size_t* indices, DistanceType* dists, size_t num_elements, bool sorted = true)
     {
     	size_t n = std::min(count_, num_elements);
     	for (size_t i=0; i<n; ++i) {
@@ -278,7 +278,7 @@ public:
      * @param num_elements Number of elements to copy
      * @param sorted Indicates if results should be sorted
      */
-    void copy(int* indices, DistanceType* dists, size_t num_elements, bool sorted = true)
+    void copy(size_t* indices, DistanceType* dists, size_t num_elements, bool sorted = true)
     {
     	size_t n = std::min(count_, num_elements);
     	for (size_t i=0; i<n; ++i) {
@@ -387,7 +387,7 @@ public:
      * @param num_elements Number of elements to copy
      * @param sorted Indicates if results should be sorted
      */
-    void copy(int* indices, DistanceType* dists, size_t num_elements, bool sorted = true)
+    void copy(size_t* indices, DistanceType* dists, size_t num_elements, bool sorted = true)
     {
     	if (sorted) {
     		// std::sort_heap(dist_index_.begin(), dist_index_.end());
@@ -489,7 +489,7 @@ public:
      * @param num_elements Number of elements to copy
      * @param sorted Indicates if results should be sorted
      */
-    void copy(int* indices, DistanceType* dists, size_t num_elements, bool sorted = true)
+    void copy(size_t* indices, DistanceType* dists, size_t num_elements, bool sorted = true)
     {
     	if (sorted) {
     		// std::sort_heap(dist_index_.begin(), dist_index_.end());
@@ -611,7 +611,7 @@ public:
      * @param num_elements Number of elements to copy
      * @param sorted Indicates if results should be sorted
      */
-    void copy(int* indices, DistanceType* dists, size_t num_elements, bool sorted = true)
+    void copy(size_t* indices, DistanceType* dists, size_t num_elements, bool sorted = true)
     {
     	if (sorted) {
     		// std::sort_heap(dist_index_.begin(), dist_index_.end());
@@ -739,7 +739,7 @@ public:
      * @param dist pointer to a C array of distances
      * @param n_neighbors the number of neighbors to copy
      */
-    void copy(int* indices, DistanceType* dist, int n_neighbors, bool sorted = true)
+    void copy(size_t* indices, DistanceType* dist, int n_neighbors, bool sorted = true)
     {
     	if (n_neighbors<0) n_neighbors = dist_indices_.size();
     	int i = 0;
